Clean up #includes and file structure in fles_ipc library
